requirements.txt
flask==2.0.1
python-dotenv==0.19.0
requests==2.26.0
anthropic>=0.18.0
httpx>=0.24.1
python-dateutil==2.8.2
gunicorn==20.1.0
werkzeug==2.2.3
google-analytics-data==0.16.2
